# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Meal {
  id: string;
  name: string;
  /** @pattern ^([01][0-9]|2[0-3]):[0-5][0-9]$ */
  time: string;
  note: string;
  /**
     * @items.minimum 0
     * @items.maximum 6
     */
  weekdays: number[];
  completed: boolean;
}

export interface MealPlan {
  id: string;
  name: string;
  /** @pattern ^([01][0-9]|2[0-3]):[0-5][0-9]$ */
  time: string;
  note: string;
  /**
     * @items.minimum 0
     * @items.maximum 6
     */
  weekdays: number[];
}

export interface MealInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
  /** @pattern ^([01][0-9]|2[0-3]):[0-5][0-9]$ */
  time: string;
  /** @maxLength 240 */
  note?: string;
  /**
     * @minItems 1
     * @items.minimum 0
     * @items.maximum 6
     */
  weekdays: number[];
}

export interface MealUpdate {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name?: string;
  /** @pattern ^([01][0-9]|2[0-3]):[0-5][0-9]$ */
  time?: string;
  /** @maxLength 240 */
  note?: string;
  /**
     * @minItems 1
     * @items.minimum 0
     * @items.maximum 6
     */
  weekdays?: number[];
}

export interface MealCompletionInput {
  date: string;
  completed: boolean;
}

export interface MealCompletion {
  mealId: string;
  date: string;
  completed: boolean;
}

export interface DashboardSummary {
  date: string;
  totalCount: number;
  doneCount: number;
  progress: number;
  nextMeal: Meal | null;
}

export interface DaySummary {
  date: string;
  totalCount: number;
  doneCount: number;
}

export interface UserPreferences {
  remindersEnabled: boolean;
  timezone: string;
  pushEnabled: boolean;
  deviceCount: number;
}

export interface PreferencesUpdate {
  remindersEnabled?: boolean;
  /**
     * @minLength 1
     * @maxLength 80
     */
  timezone?: string;
}

export interface PushPublicKey {
  publicKey: string;
  enabled: boolean;
}

export interface PushStatus {
  pushEnabled: boolean;
  deviceCount: number;
}

export interface PushSubscriptionKeys {
  p256dh: string;
  auth: string;
}

export interface PushSubscriptionInput {
  /** @maxLength 2048 */
  endpoint: string;
  /** @nullable */
  expirationTime?: number | null;
  keys: PushSubscriptionKeys;
}

export interface PushSubscriptionDelete {
  /** @maxLength 2048 */
  endpoint: string;
}

export interface PushTestResult {
  sent: boolean;
  deviceCount: number;
  message: string;
}

export type ListMealsParams = {
date: string;
};

export type GetDashboardSummaryParams = {
date: string;
};

export type GetJourneySummaryParams = {
startDate: string;
endDate: string;
};