# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Meal } from './meal';

export interface DashboardSummary {
  date: Date;
  totalCount: number;
  doneCount: number;
  progress: number;
  nextMeal: Meal | null;
}