# lib/api-zod/src/generated/types/mealInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MealInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
  /** @pattern ^([01][0-9]|2[0-3]):[0-5][0-9]$ */
  time: string;
  /** @maxLength 240 */
  note?: string;
  /**
     * @minItems 1
     * @items.minimum 0
     * @items.maximum 6
     */
  weekdays: number[];
}