# lib/api-zod/src/generated/types/mealPlan.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MealPlan {
  id: string;
  name: string;
  /** @pattern ^([01][0-9]|2[0-3]):[0-5][0-9]$ */
  time: string;
  note: string;
  /**
     * @items.minimum 0
     * @items.maximum 6
     */
  weekdays: number[];
}