# lib/api-zod/src/generated/types/pushSubscriptionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PushSubscriptionKeys } from './pushSubscriptionKeys';

export interface PushSubscriptionInput {
  /** @maxLength 2048 */
  endpoint: string;
  /** @nullable */
  expirationTime?: number | null;
  keys: PushSubscriptionKeys;
}