# lib/api-zod/src/generated/types/pushTestResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PushTestResult {
  sent: boolean;
  deviceCount: number;
  message: string;
}